Testing Update
Testing Report 1
Did a couple more tests on compatible images
most images that go through have 32 bit depth and are pngs, havent
tested it for jpegs yet. Does it have something to do with transparency?

Testing Report 2
The UV coordinates seem to be okay for .obj models, although the
coordinates seem to be flipped. I had to flip the img src vertically so
itd fit the model... or.. read the textureUVs backwards?