Rearranging Framework
Attempted to move addition of gameobjects to testbed:

- Added a list of gameobjects in Scene
- Was unable to do OpenGLRenderer::render(Scene* .. ) due to circular
dependency, switched to OpenGLRenderer::render(GameObject* ..) instead
- OpenGLRenderer::render contains temporary implementation until a more
permanent solution is discussed.

Notes: Models are still being pushed to the GPU on loadModel("...") as
opposed to having them all pushed at the same time in Scene::render().
This is because it requires DrawData to contain the actual vertices that
will be merged with other drawData from different objects later on.