Testing normals
The crate's normals are just weird, maybe the creator just did it for
effect, its more spherical than boxy.

Added a commentable line in fragshader.
enable fColor = vNormal to check the normals of the objects.

TODO next for me: fix ModelManager load to return an ID when the model
is already in the map container